Refactored the plugin code.
Change the directory structure to be as the needed directory structure
of a vim plugin, and refactor the code.
When refactoring the code, I mainly moved the functions to autoload
directory, and set the namespace for them. I did little changes to the
functions themselves, and most of them were TODO fixes.

I didn't add or remove any functionality from the code itself.