Add handling for empty store keys
Added special handling for the case in which the store keys would be
empty. In the previous version, this would have raised an error message
when openning vim. Now, I added a test to validate that they are not
empty, if they would be empty the lines would not run, and therefore,
won't raise any error message.

I also added a flag to allow the user decide if he wants to use those
mapping or not, to allow more control over those variables, without
explicitly setting them to empty strings.

This should solve the issue #5