Change the doc file to generate better tags
The previous tags generation did not generate enough tags for the
variables. The tags that was generated was only with prefix of g: for
all the variables. This make jumping between the tags in the help file
hard.

To fix it, I added some more tags in the help file, to make it easier
to jump inside this file.