Add list choosing in case of multiple file matches
Added a new option to choose between macro files in case there are more
than one match for the glob search. Until now, the behavior in this case
was unexpected, since glob returned all the values one after the other,
and the behavior was different between computers.
Now, glob would always return a list with all the matching values. In
case there would be more than one match in this list, the user would be
asked to choose which file he wants to run. The regular behavior of the
plugin (when there is only one result) stays the same.

Now, the use of wildcards in the search can be used more wildly, since
it would let the user choose the wanted macro in case of wildcard that
matches more than one file. So, I added a documentation about it as
a tip, both in the readme file and in the plugin doc directory.

This commit solves issue #6