Update the installation method in readme
The previous installation method in the readme was correct when the
plugin had only a single file. Since now the plugin has many files in
the plugin's different directories, the installation is a bit different.

Since the installation is now can be a bit more complicated, it is
different for the version of vim the user is running and the importance
of loading the different files as he sees it.

This solves issue #11