Handle special paths in marvim_store
In the case that there are special character in the marvim_store
variable, they weren't expended, and the paths turn out to be different
than the wanted paths.

To fix this bug, a call to expand was added, so the path will be
expanded to the wanted path before using it.

Now, the plugin handle the case of having ~ for example in the value of
g:marvim_store, and it will get the wanted path in the home directory.

Partially fixes #13